!6 add compiled cami bytecode * add compiled cami bytecode * amend test cases * add command line arguments * add useful tools in etc directory

+++ evaluation/test_suite/positive/ub24/1.c
@@ -4,8 +4,8 @@
 int main()
 {
     int16_t x[4] = {0};
-    int32_t* p = (int16_t*)&x[0];
-    int32_t* q = (int16_t*)&x[1];
+    int32_t* p = (int32_t*)&x[0];
+    int32_t* q = (int32_t*)&x[1];
     printf("%d\n", (int)*(char*)p);
     printf("%d\n", (int)*(char*)q);
 }
